+++ src/timer.h
@@ -1,12 +1,12 @@
 #ifndef TIMER_H
 #define TIMER_H
 
-int timer_init(double ticks_per_sec);
+int timer_init(float ticks_per_sec);
 void timer_exit(void);
-double timer_ticks_per_sec(void);
+float timer_ticks_per_sec(void);
 unsigned int timer_get_ticks(void);
 unsigned int timer_get_ticks_delta(void);
-double timer_get_time(void);
-double timer_get_time_delta(void);
+float timer_get_time(void);
+float timer_get_time_delta(void);
 
 #endif
